refactor: modularize solver and utilities
