+++ lpm.py
@@ -24,23 +24,9 @@ from typing import Dict, List, Optional, Set, Tuple, Iterable
 from collections import deque
 import zstandard as zstd
 
-# =========================== Config / Defaults ================================
-CONF_FILE = Path("/etc/lpm/lpm.conf")      # KEY=VALUE, e.g. ARCH=znver2
-STATE_DIR = Path(os.environ.get("LPM_STATE_DIR", "/var/lib/lpm"))
-DB_PATH   = STATE_DIR / "state.db"
-CACHE_DIR = STATE_DIR / "cache"
-SNAPSHOT_DIR = STATE_DIR / "snapshots"
-REPO_LIST = STATE_DIR / "repos.json"       # [{"name":"core","url":"file:///srv/repo","priority":10}, ...]
-PIN_FILE  = STATE_DIR / "pins.json"        # {"hold":["pkg"], "prefer":{"pkg":"~=3.3"}}
-HOOK_DIR  = Path("/usr/share/lpm/hooks")
-SIGN_KEY  = Path("/etc/lpm/private/lpm_signing.pem")   # OpenSSL PEM private key for signing
-TRUST_DIR = Path("/etc/lpm/trust")                     # dir of *.pem public keys for verification
-DEFAULT_ROOT = "/"
-UMASK = 0o22
-os.umask(UMASK)
-for d in (STATE_DIR, CACHE_DIR, SNAPSHOT_DIR): d.mkdir(parents=True, exist_ok=True)
-if not REPO_LIST.exists(): REPO_LIST.write_text("[]", encoding="utf-8")
-if not PIN_FILE.exists(): PIN_FILE.write_text(json.dumps({"hold":[],"prefer":{}}, indent=2), encoding="utf-8")
+from src.config import *  # noqa: F401,F403
+from src.fs import read_json, write_json, urlread
+from src.solver import CNF, CDCLSolver, cdcl_solve
 
 # =========================== Protected packages ===============================
 PROTECTED_FILE = Path("/etc/lpm/protected.json")
@@ -87,110 +73,6 @@ from tqdm import tqdm
 def progress_bar(iterable, desc="Processing", unit="item"):
     return tqdm(iterable, desc=desc, unit=unit, ncols=80, colour="cyan")
 
-# =========================== JSON / Config ====================================
-def read_json(p: Path):
-    with p.open("r", encoding="utf-8") as f: return json.load(f)
-def write_json(p: Path, obj):
-    tmp = p.with_suffix(".tmp")
-    with tmp.open("w", encoding="utf-8") as f: json.dump(obj, f, indent=2, sort_keys=True)
-    tmp.replace(p)
-
-def urlread(url: str) -> bytes:
-    with urllib.request.urlopen(url) as r:
-        total = int(r.headers.get("content-length", 0) or 0)
-        if total == 0:
-            return r.read()
-        chunk_size = 1 << 14
-        data = bytearray()
-        with tqdm(total=total, desc="Downloading", unit="B", unit_scale=True, ncols=80, colour="cyan") as bar:
-            while True:
-                chunk = r.read(chunk_size)
-                if not chunk: break
-                data.extend(chunk)
-                bar.update(len(chunk))
-        return bytes(data)
-
-def load_conf(path: Path) -> Dict[str,str]:
-    if not path.exists(): return {}
-    out = {}
-    for ln in path.read_text(encoding="utf-8").splitlines():
-        ln = ln.strip()
-        if not ln or ln.startswith("#"): continue
-        if "=" in ln:
-            k, v = ln.split("=",1)
-            out[k.strip()] = v.strip()
-    return out
-    
-# ================ CONFIG LOADER:: /etc/lpm/lpm.conf ====
-CONF = load_conf(CONF_FILE)
-ARCH = CONF.get("ARCH", os.uname().machine if hasattr(os, "uname") else "x86_64")
-
-# --- Optimization level (-O2 etc.) ---
-OPT_LEVEL = CONF.get("OPT_LEVEL", "-O2")
-if OPT_LEVEL not in ("-Os", "-O2", "-O3", "-Ofast"):
-    OPT_LEVEL = "-O2"
-
-# --- Snapshot retention ---
-try:
-    MAX_SNAPSHOTS = max(0, int(CONF.get("MAX_SNAPSHOTS", "10")))
-except ValueError:
-    MAX_SNAPSHOTS = 10
-
-
-def _detect_cpu() -> Tuple[str, str, str, str]:
-    """Return (march, mtune, vendor, family)."""
-    vendor = family = ""  # defaults
-    try:
-        with open("/proc/cpuinfo", "r", encoding="utf-8") as f:
-            for line in f:
-                if not vendor and line.startswith("vendor_id"):
-                    vendor = line.split(":", 1)[1].strip()
-                elif not family and line.startswith("cpu family"):
-                    family = line.split(":", 1)[1].strip()
-                if vendor and family:
-                    break
-    except Exception:
-        pass
-
-    march = mtune = "generic"
-    try:
-        fam = int(family)
-    except Exception:
-        fam = None
-
-    if vendor == "AuthenticAMD":
-        if fam and fam >= 25:
-            march = mtune = "znver4"
-        elif fam and fam >= 24:
-            march = mtune = "znver3"
-        elif fam and fam >= 23:
-            march = mtune = "znver2"
-    elif vendor == "GenuineIntel":
-        if fam and fam >= 6:
-            march = mtune = "x86-64-v3"
-
-    return march, mtune, vendor, family
-
-
-MARCH, MTUNE, CPU_VENDOR, CPU_FAMILY = _detect_cpu()
-
-# ================ Init System Detection ===============================================
-def detect_init_system() -> str:
-    """
-    Detect which init system is active.
-    Returns: 'systemd', 'runit', 'openrc', 'sysv', or 'unknown'
-    """
-    if shutil.which("systemctl") and os.path.isdir("/run/systemd/system"):
-        return "systemd"
-    if os.path.isdir("/etc/runit") or os.path.isdir("/etc/runit/runsvdir"):
-        return "runit"
-    if os.path.isdir("/etc/init.d"):
-        if shutil.which("openrc"):
-            return "openrc"
-        return "sysv"
-    return "unknown"
-    
-    
 # ============================ Build Isolation =======================
 def sandboxed_run(func: str, cwd: Path, env: dict, script_path: Path, stagedir: Path, buildroot: Path, srcroot: Path):
     """
@@ -539,322 +421,6 @@ def providers_for(u: Universe, atom: Atom) -> List[PkgMeta]:
     if atom.op and atom.ver:
         cands = [p for p in cands if satisfies(p.version, f"{atom.op}{atom.ver}")]
     return cands
-
-# =========================== CNF / SAT ========================================
-class CNF:
-    """Simple CNF container with watched literal management."""
-
-    def __init__(self):
-        self.clauses: List[List[int]] = []
-        self.next_var = 1
-        self.varname: Dict[int, str] = {}
-        self.namevar: Dict[str, int] = {}
-        # Watched literal data structures
-        self.watch_list: Dict[int, List[int]] = {}
-        self.watchers: List[Tuple[int, int]] = []
-        # Clause activity for database reduction
-        self.activity: List[float] = []
-        # Indices of learnt clauses
-        self.learnts: Set[int] = set()
-
-    def new_var(self, name: str) -> int:
-        if name in self.namevar:
-            return self.namevar[name]
-        v = self.next_var
-        self.next_var += 1
-        self.namevar[name] = v
-        self.varname[v] = name
-        return v
-
-    def add_clause(self, clause: List[int], learnt: bool = False) -> int:
-        idx = len(self.clauses)
-        self.clauses.append(clause)
-        self.activity.append(0.0)
-        if learnt:
-            self.learnts.add(idx)
-        if not clause:
-            self.watchers.append((0, 0))
-            return idx
-        if len(clause) == 1:
-            lit = clause[0]
-            self.watchers.append((lit, lit))
-            self.watch_list.setdefault(lit, []).append(idx)
-        else:
-            a, b = clause[0], clause[1]
-            self.watchers.append((a, b))
-            self.watch_list.setdefault(a, []).append(idx)
-            self.watch_list.setdefault(b, []).append(idx)
-        return idx
-
-    def add(self, *clauses: Iterable[int]):
-        for c in clauses:
-            clause = list(c)
-            if clause:
-                self.add_clause(clause)
-
-    def remove_clause(self, idx: int) -> None:
-        clause = self.clauses[idx]
-        if not clause:
-            return
-        w1, w2 = self.watchers[idx]
-        if w1 in self.watch_list and idx in self.watch_list[w1]:
-            self.watch_list[w1].remove(idx)
-            if not self.watch_list[w1]:
-                del self.watch_list[w1]
-        if w2 in self.watch_list and idx in self.watch_list[w2]:
-            self.watch_list[w2].remove(idx)
-            if not self.watch_list[w2]:
-                del self.watch_list[w2]
-        self.clauses[idx] = []
-        self.watchers[idx] = (0, 0)
-        self.activity[idx] = 0.0
-        self.learnts.discard(idx)
-
-
-class SATResult:
-    def __init__(self, sat: bool, assign: Dict[int, bool]):
-        self.sat = sat
-        self.assign = assign
-
-
-@dataclass
-class Implication:
-    """Node in the implication graph."""
-    level: int = 0
-    reason: Optional[int] = None
-    preds: List[int] = field(default_factory=list)
-
-
-def luby(i: int) -> int:
-    """Return the i-th value of the Luby sequence."""
-    k = 1
-    while (1 << k) - 1 < i:
-        k += 1
-    if i == (1 << k) - 1:
-        return 1 << (k - 1)
-    return luby(i - (1 << (k - 1)) + 1)
-
-
-def cdcl_solve(cnf: CNF, prefer_true: Set[int], prefer_false: Set[int],
-               bias: Optional[Dict[int,float]] = None,
-               decay_map: Optional[Dict[int,float]] = None) -> SATResult:
-    """CDCL SAT solver with non-chronological backtracking and restarts."""
-
-    nvars = cnf.next_var - 1
-    assigns: Dict[int, Optional[bool]] = {i: None for i in range(1, nvars + 1)}
-    levels: Dict[int, int] = {i: 0 for i in range(1, nvars + 1)}
-    reason: Dict[int, Optional[int]] = {i: None for i in range(1, nvars + 1)}
-    trail: List[int] = []
-    trail_lim: List[int] = []
-    queue = deque()
-    # Implication graph nodes for each variable
-    imp_graph: Dict[int, Implication] = {i: Implication() for i in range(1, nvars + 1)}
-
-    # Variable activity and phase saving
-    bias = bias or {}
-    decay_map = decay_map or {}
-    var_activity: Dict[int,float] = {i: bias.get(i, 0.0) for i in range(1, nvars + 1)}
-    saved_phase: Dict[int,bool] = {}
-    var_inc = 1.0
-    var_decay_conf = float(CONF.get("VSIDS_VAR_DECAY", "0.95"))
-    def bump_var(v: int):
-        var_activity[v] += var_inc
-        if var_activity[v] > 1e100:
-            for k in var_activity:
-                var_activity[k] *= 1e-100
-            var_inc *= 1e-100
-    def decay_var_activity():
-        nonlocal var_inc
-        var_inc /= var_decay_conf
-        for v in var_activity:
-            factor = decay_map.get(v, var_decay_conf)
-            var_activity[v] *= factor
-
-    # Clause activity management
-    cla_inc = 1.0
-    cla_decay = float(CONF.get("VSIDS_CLAUSE_DECAY", "0.999"))
-
-    def bump_clause(ci: Optional[int]):
-        if ci is not None:
-            cnf.activity[ci] += cla_inc
-
-    def decay_clause_activity():
-        nonlocal cla_inc
-        cla_inc /= cla_decay
-
-    def reduce_db():
-        learnts = [idx for idx in cnf.learnts if cnf.clauses[idx]]
-        if len(learnts) <= 20:
-            return
-        learnts.sort(key=lambda idx: cnf.activity[idx])
-        reasons = set(reason.values())
-        for idx in learnts[:len(learnts)//2]:
-            if idx not in reasons and len(cnf.clauses[idx]) > 2:
-                cnf.remove_clause(idx)
-
-    def current_level() -> int:
-        return len(trail_lim)
-
-    def value(lit: int) -> Optional[bool]:
-        val = assigns[abs(lit)]
-        if val is None:
-            return None
-        return val if lit > 0 else not val
-
-    def enqueue(lit: int, rsn: Optional[int]) -> None:
-        v = abs(lit)
-        val = lit > 0
-        if assigns[v] is not None:
-            return
-        assigns[v] = val
-        saved_phase[v] = val
-        levels[v] = current_level()
-        reason[v] = rsn
-        trail.append(lit)
-        queue.append(lit)
-        node = imp_graph[v]
-        node.level = levels[v]
-        node.reason = rsn
-        if rsn is not None:
-            node.preds = [l for l in cnf.clauses[rsn] if abs(l) != v]
-        else:
-            node.preds = []
-
-    # Initialize with unit clauses
-    for i, cl in enumerate(cnf.clauses):
-        if len(cl) == 1:
-            enqueue(cl[0], i)
-
-    def propagate() -> Optional[int]:
-        while queue:
-            lit = queue.popleft()
-            for ci in list(cnf.watch_list.get(-lit, [])):
-                clause = cnf.clauses[ci]
-                w1, w2 = cnf.watchers[ci]
-                if w1 == -lit:
-                    other = w2
-                    first = True
-                else:
-                    other = w1
-                    first = False
-                if value(other) is True:
-                    continue
-                found = False
-                for new_lit in clause:
-                    if new_lit == other or new_lit == -lit:
-                        continue
-                    if value(new_lit) is not False:
-                        if first:
-                            cnf.watchers[ci] = (new_lit, other)
-                        else:
-                            cnf.watchers[ci] = (other, new_lit)
-                        cnf.watch_list[-lit].remove(ci)
-                        cnf.watch_list.setdefault(new_lit, []).append(ci)
-                        found = True
-                        break
-                if not found:
-                    if value(other) is False:
-                        return ci
-                    else:
-                        enqueue(other, ci)
-        return None
-
-    def pick_branch_var() -> int:
-        unassigned = [v for v in range(1, nvars + 1) if assigns[v] is None]
-        if not unassigned:
-            return 0
-        return max(unassigned, key=lambda v: var_activity[v])
-
-    def analyze(conflict_idx: int) -> Tuple[List[int], int]:
-        bump_clause(conflict_idx)
-        for lit in cnf.clauses[conflict_idx]:
-            bump_var(abs(lit))
-        seen: Set[int] = set()
-        learnt: List[int] = []
-        counter = 0
-        clause = cnf.clauses[conflict_idx][:]
-        i = len(trail) - 1
-        while True:
-            for lit in clause:
-                v = abs(lit)
-                bump_var(v)
-                node = imp_graph[v]
-                if v not in seen and node.level > 0:
-                    seen.add(v)
-                    if node.level == current_level():
-                        counter += 1
-                    else:
-                        learnt.append(lit)
-            while True:
-                lit = trail[i]
-                i -= 1
-                if abs(lit) in seen:
-                    break
-            v = abs(lit)
-            clause_idx = imp_graph[v].reason
-            bump_clause(clause_idx)
-            if clause_idx is not None:
-                for l in cnf.clauses[clause_idx]:
-                    bump_var(abs(l))
-            clause = imp_graph[v].preds.copy() if clause_idx is not None else []
-            counter -= 1
-            if counter <= 0:
-                learnt.append(-lit)
-                break
-        back_lvl = 0
-        if len(learnt) > 1:
-            back_lvl = max(imp_graph[abs(l)].level for l in learnt[:-1])
-        for lit in learnt:
-            bump_var(abs(lit))
-        return learnt, back_lvl
-
-    def backtrack(level: int) -> None:
-        while current_level() > level:
-            start = trail_lim.pop()
-            while len(trail) > start:
-                lit = trail.pop()
-                v = abs(lit)
-                assigns[v] = None
-                reason[v] = None
-                levels[v] = 0
-                imp_graph[v] = Implication()
-            queue.clear()
-
-    conflicts = 0
-    restart_count = 1
-    restart_limit = luby(restart_count) * 100
-
-    while True:
-        confl = propagate()
-        if confl is not None:
-            conflicts += 1
-            if current_level() == 0:
-                return SATResult(False, {v: False for v in assigns})
-            learnt, back_lvl = analyze(confl)
-            ci = cnf.add_clause(learnt, learnt=True)
-            bump_clause(ci)
-            backtrack(back_lvl)
-            enqueue(learnt[0], ci)
-            decay_clause_activity()
-            decay_var_activity()
-            if conflicts % 200 == 0:
-                reduce_db()
-            if conflicts >= restart_limit:
-                restart_count += 1
-                restart_limit = luby(restart_count) * 100
-                backtrack(0)
-        else:
-            v = pick_branch_var()
-            if v == 0:
-                final = {var: (assigns[var] if assigns[var] is not None else False) for var in assigns}
-                return SATResult(True, final)
-            trail_lim.append(len(trail))
-            phase = saved_phase.get(v)
-            if phase is None:
-                lit = -v if v in prefer_false and v not in prefer_true else v
-            else:
-                lit = v if phase else -v
-            enqueue(lit, None)
 
 # =========================== Resolver encoding =================================
 def expr_to_cnf_disj(u: Universe, e: DepExpr, cnf: CNF, var_of: Dict[Tuple[str,str],int]) -> List[int]:
